# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * EMPOWERMENT Theater Ticket Reservation System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Reservation {
  id: number;
  name: string;
  email: string;
  ticketCount: number;
  /** @nullable */
  specialNeeds?: string | null;
  token: string;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  createdAt: string;
}

export interface ReservationInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  ticketCount: number;
  specialNeeds?: string;
}

export interface CheckInInput {
  adminKey: string;
}

export interface AdminKeyInput {
  adminKey: string;
}

export interface ReservationUpdate {
  adminKey: string;
  /** @minLength 2 */
  name?: string;
  email?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  ticketCount?: number;
  /** @nullable */
  specialNeeds?: string | null;
  checkedIn?: boolean;
}

export interface Stats {
  totalReservations: number;
  totalTickets: number;
  totalCheckedIn: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListReservationsParams = {
adminKey: string;
};