# lib/api-zod/src/generated/types/checkInInput.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * EMPOWERMENT Theater Ticket Reservation System
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInInput {
  adminKey: string;
}