# lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * EMPOWERMENT Theater Ticket Reservation System
 * OpenAPI spec version: 0.1.0
 */

export interface Reservation {
  id: number;
  name: string;
  email: string;
  ticketCount: number;
  /** @nullable */
  specialNeeds?: string | null;
  token: string;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  createdAt: string;
}