# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * EMPOWERMENT Theater Ticket Reservation System
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  ticketCount: number;
  specialNeeds?: string;
}