# lib/api-zod/src/generated/types/reservationUpdate.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * EMPOWERMENT Theater Ticket Reservation System
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationUpdate {
  adminKey: string;
  /** @minLength 2 */
  name?: string;
  email?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  ticketCount?: number;
  /** @nullable */
  specialNeeds?: string | null;
  checkedIn?: boolean;
}